requirements.txt
numpy<2
matplotlib
Pillow
opencv-python
argcomplete
pycocotools==2.0.8
wandb
tensorboard
pandas
scipy<=1.7.3
tqdm
lap
addict
motmetrics
terminaltables
seaborn
scalabel
torch
torchvision

# MM Ecosystem
mmcv-full==1.7.2
mmdet==2.28.2
mmtrack==0.14.0

git+https://github.com/bdd100k/bdd100k.git
git+https://github.com/JonathonLuiten/TrackEval.git

# For RCBEVDet
timm>=0.6.0
einops>=0.6.0

# Optional: NuScenes dataset adapter support
nuscenes-devkit

# For visualization purposes
open3d
napari[pyqt5]
plotly

